requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.8.0
scipy>=1.10.0
tqdm>=4.65.0

# Deep learning frameworks
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Training utilities
wandb>=0.15.0
lpips>=0.1.4

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Video processing
moviepy>=1.0.3
scenedetect[opencv]>=0.6.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Optional (for advanced features)
# peft>=0.4.0  # For LoRA
# xformers>=0.0.20  # For memory-efficient attention
